@@ -3,17 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.21.0",
-    "moment": "^2.24.0",
-    "react": "^17.0.0",
-    "react-app-rewired": "^2.1.3",
-    "react-dom": "^17.0.0",
-    "react-helmet": "^6.0.0",
-    "react-redux": "^7.0.3",
-    "react-router": "^6.28.0",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "^4.0.0",
-    "redux": "^4.0.1"
+    "axios": "^1.9.0",
+    "moment": "^2.30.1",
+    "react": "^19.1.0",
+    "react-app-rewired": "^2.2.1",
+    "react-dom": "^19.1.0",
+    "react-helmet": "^6.1.0",
+    "react-redux": "^9.2.0",
+    "react-router": "^7.6.1",
+    "react-router-dom": "^7.6.1",
+    "react-scripts": "^5.0.1",
+    "redux": "^5.0.1"
   },
   "scripts": {
     "build-dev": "NODE_OPTIONS=--openssl-legacy-provider yarn exec dotenv -e .env.development -- yarn exec react-app-rewired build",
@@ -37,9 +37,11 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/user-event": "^14.6.1",
     "dotenv-cli": "^8.0.0",
-    "eslint": "^7.32.0",
-    "eslint-config-react-app": "^7.0.1",
+    "eslint": "^8.57.0",
+    "eslint-config-react-app": "^7.1.0-next.26",
     "jest-environment-jsdom": "30.0.0-beta.3",
+    "typescript": "^5.8.3",
+    "webpack": "^5.99.9",
     "workbox-webpack-plugin": "^7.3.0"
   }
 }
