@@ -1,16 +1,9 @@
 import React from "react";
-import { Route, Redirect } from "react-router-dom";
+import { Navigate } from "react-router-dom";
 import { withAuth } from "../lib/AuthProvider";
 
-function PrivateRoute({ component: Component, isLoggedin, ...rest }) {
-  return (
-    <Route
-      {...rest}
-      render={props =>
-        isLoggedin ? <Component {...props} /> : <Redirect to="/login" />
-      }
-    />
-  );
+function PrivateRoute({ children, isLoggedin }) {
+  return isLoggedin ? children : <Navigate to="/login" replace />;
 }
 
 export default withAuth(PrivateRoute);
