@@ -87,7 +87,7 @@ function Pantry(props) {
       })
       .catch((error) => console.log(error))
 
-  }, [])
+  }, [props.user._id])
 
   useEffect(() => {
     if (myRef.current) {
@@ -127,11 +127,11 @@ function Pantry(props) {
       </div>
       <div className="form-row">
         <div className="col">
-          <button className="btn btn-primary" type="submit" onClick={handleSearchByIngredients} disabled={disabled}><span className="badge badge-light">{selectedIngredients.length}</span> ingredient{selectedIngredients.length !== 1 ? 's' : null } selected</button>
+          <button className="btn btn-primary" type="submit" onClick={handleSearchByIngredients} disabled={disabled}><span className="badge badge-light">{selectedIngredients.length}</span> ingredient{selectedIngredients.length !== 1 ? 's' : null} selected</button>
         </div>
       </div>
       {performSearch ?
-        <h2>{recipes.length} recipe{recipes.length !== 1 ? 's use the selected ingredients' : ' uses the selected ingredient' }</h2>
+        <h2>{recipes.length} recipe{recipes.length !== 1 ? 's use the selected ingredients' : ' uses the selected ingredient'}</h2>
         : null}
       {recipes.map((recipe) =>
         <div className="card mb-3" key={recipe._id._id}>
@@ -144,7 +144,7 @@ function Pantry(props) {
                 <h5 className="card-title">
                   <Link to={{
                     pathname: `/recipes/${recipe._id._id}`
-                  }}>{recipe._id.name}</Link> <span className="badge badge-info">{recipe.matches} ingredient{recipe.matches > 1 ? 's' : null } matched</span>
+                  }}>{recipe._id.name}</Link> <span className="badge badge-info">{recipe.matches} ingredient{recipe.matches > 1 ? 's' : null} matched</span>
                 </h5>
                 <p className="card-text">{recipe._id.description}</p>
               </div>
